move isSupervisor check to shift details page
